export generateTranslatableConfig
