+++ packages/visual-editor/src/internal/puck/constant-value-fields/ProductSection.tsx
@@ -7,7 +7,7 @@ import {
 } from "../../../types/types.ts";
 import { translatableCTAFields } from "./CallToAction.tsx";
 import { usePlatformTranslation } from "../../../utils/i18nPlatform.ts";
-import { generateTranslatableConstantConfig } from "./Text.tsx";
+import { generateTranslatableConfig } from "../../../utils/generateTranslatableConfig.tsx";
 import { resolveTranslatableString } from "../../../utils/resolveTranslatableString.tsx";
 import { useMemo } from "react";
 
@@ -39,7 +39,7 @@ const ProductStructArrayField = (): ArrayField<ProductStruct[]> => {
   const { t, i18n } = usePlatformTranslation();
 
   const nameField = useMemo(() => {
-    return generateTranslatableConstantConfig<TranslatableString | undefined>(
+    return generateTranslatableConfig<TranslatableString | undefined>(
       {
         key: "name",
         options: {
@@ -51,7 +51,7 @@ const ProductStructArrayField = (): ArrayField<ProductStruct[]> => {
   }, []);
 
   const categoryField = useMemo(() => {
-    return generateTranslatableConstantConfig<TranslatableString | undefined>(
+    return generateTranslatableConfig<TranslatableString | undefined>(
       {
         key: "category",
         options: {
@@ -63,7 +63,7 @@ const ProductStructArrayField = (): ArrayField<ProductStruct[]> => {
   }, []);
 
   const descriptionField = useMemo(() => {
-    return generateTranslatableConstantConfig<TranslatableRTF2 | undefined>(
+    return generateTranslatableConfig<TranslatableRTF2 | undefined>(
       {
         key: "description",
         options: {
