+++ packages/visual-editor/src/internal/puck/constant-value-fields/Text.tsx
@@ -1,10 +1,6 @@
-import { AutoField, CustomField, FieldLabel, TextField } from "@measured/puck";
-import React from "react";
+import { CustomField, TextField } from "@measured/puck";
 import { TranslatableRTF2, TranslatableString } from "../../../types/types.ts";
-import { useDocument } from "../../../hooks/useDocument.tsx";
-import { getDisplayValue } from "../../../utils/resolveTranslatableString.tsx";
-import { usePlatformTranslation } from "@yext/visual-editor";
-import { Translation } from "../../types/translation.ts";
+import { generateTranslatableConfig } from "../../../utils/generateTranslatableConfig.tsx";
 
 export const TEXT_CONSTANT_CONFIG: TextField = {
   type: "text",
@@ -12,44 +8,7 @@ export const TEXT_CONSTANT_CONFIG: TextField = {
 };
 
 export const TRANSLATABLE_STRING_CONSTANT_CONFIG: CustomField<TranslatableString> =
-  generateTranslatableConstantConfig<TranslatableString>(undefined, "text");
+  generateTranslatableConfig<TranslatableString>(undefined, "text");
 
 export const TRANSLATABLE_RTF2_CONSTANT_CONFIG: CustomField<TranslatableRTF2> =
-  generateTranslatableConstantConfig<TranslatableRTF2>(undefined, "text");
-
-export function generateTranslatableConstantConfig<
-  T extends TranslatableRTF2 | undefined,
->(label?: Translation, fieldType?: "text" | "textarea"): CustomField<T> {
-  return {
-    type: "custom",
-    render: ({ onChange, value }) => {
-      const document: any = useDocument();
-      const locale = document?.locale ?? "en";
-      const { t } = usePlatformTranslation();
-      const autoField = (
-        <AutoField
-          field={{ type: fieldType ?? "text" }}
-          value={getDisplayValue(value, locale)}
-          onChange={(val) =>
-            onChange({
-              ...(typeof value === "object" && !Array.isArray(value)
-                ? value
-                : {}),
-              [locale]: val,
-            } as T)
-          }
-        />
-      );
-
-      if (!label) {
-        return <div className={"ve-pt-3"}>{autoField}</div>;
-      }
-
-      return (
-        <FieldLabel label={t(label.key, label.options) + ` (${locale})`}>
-          {autoField}
-        </FieldLabel>
-      );
-    },
-  };
-}
+  generateTranslatableConfig<TranslatableRTF2>(undefined, "text");
