+++ packages/visual-editor/src/utils/index.ts
@@ -48,3 +48,4 @@ export {
   msg,
   pt,
 } from "./i18nPlatform.ts";
+export { generateTranslatableConfig } from "./generateTranslatableConfig.tsx";
